Refactor: Distribute endpoints and services by entities
