+++ api/test/e2e/admin-regions/admin-regions-get-trees.spec.ts
@@ -5,11 +5,16 @@ import { AppModule } from 'app.module';
 import { AdminRegion } from 'modules/admin-regions/admin-region.entity';
 import { AdminRegionsModule } from 'modules/admin-regions/admin-regions.module';
 import { AdminRegionRepository } from 'modules/admin-regions/admin-region.repository';
-import { createH3Data, createAdminRegion } from '../../entity-mocks';
+import {
+  createH3Data,
+  createAdminRegion,
+  createSourcingLocation,
+} from '../../entity-mocks';
 import { H3DataRepository } from 'modules/h3-data/h3-data.repository';
 import { expectedJSONAPIAttributes } from './config';
 import { saveUserAndGetToken } from '../../utils/userAuth';
 import { getApp } from '../../utils/getApp';
+import { Material } from '../../../src/modules/materials/material.entity';
 
 //TODO: Allow these tests when feature fix is merged
 describe('AdminRegions - Get trees', () => {
@@ -222,4 +227,53 @@ describe('AdminRegions - Get trees', () => {
       },
     );
   });
+  test(
+    'When I query a Impact Admin-Region Tree endpoint' +
+      'Then I should receive a tree list of admin imported where sourcing locations of the user are' +
+      'And if any of them is a child admin-region, I should get its parent too',
+    async () => {
+      const parentAdminRegion: AdminRegion = await createAdminRegion({
+        name: 'parentAdminRegion',
+      });
+      const childAdminRegion: AdminRegion = await createAdminRegion({
+        name: 'childAdminRegion',
+        parent: parentAdminRegion,
+      });
+
+      const parentWithNoChildAdminRegion: AdminRegion = await createAdminRegion(
+        {
+          name: 'parentWithNoChild',
+        },
+      );
+      const adminRegionNotPresentInSourcingLocations: AdminRegion =
+        await createAdminRegion({
+          name: 'adminRegionNotPresentInSourcingLocations',
+        });
+
+      for await (const adminRegion of [
+        childAdminRegion,
+        parentWithNoChildAdminRegion,
+      ]) {
+        await createSourcingLocation({ adminRegionId: adminRegion.id });
+      }
+
+      const response = await request(app.getHttpServer())
+        .get('/api/v1/admin-regions/trees/user')
+        .set('Authorization', `Bearer ${jwtToken}`)
+        .expect(HttpStatus.OK);
+
+      expect(response.body.data[0].id).toEqual(parentWithNoChildAdminRegion.id);
+      expect(response.body.data[0].attributes.children).toEqual([]);
+      expect(response.body.data[1].id).toEqual(parentAdminRegion.id);
+      expect(response.body.data[1].attributes.children[0].id).toEqual(
+        childAdminRegion.id,
+      );
+      expect(
+        response.body.data.find(
+          (material: Material) =>
+            material.id === adminRegionNotPresentInSourcingLocations.id,
+        ),
+      ).toBe(undefined);
+    },
+  );
 });
